# default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for my project",
    "version": "1.0.0",
    "title": "My API Documentation",
    "contact": {
      "name": "John Doe",
      "url": "http://example.com",
      "email": "john@example.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/admin/log/page": {
      "get": {
        "tags": [
          "日志管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "operationType",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "operatorName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«LogPageVO»",
              "originalRef": "PageDTO«LogPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/admin/notification/create": {
      "post": {
        "tags": [
          "通知管理"
        ],
        "summary": "create",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "notificationDTO",
            "description": "notificationDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NotificationDTO",
              "originalRef": "NotificationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/notification/delete": {
      "delete": {
        "tags": [
          "通知管理"
        ],
        "summary": "deleteNotification",
        "operationId": "deleteNotificationUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/notification/detail/{id}": {
      "get": {
        "tags": [
          "通知管理"
        ],
        "summary": "queryNotificationDetail",
        "operationId": "queryNotificationDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/NotificationDetailVO",
              "originalRef": "NotificationDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/notification/page": {
      "get": {
        "tags": [
          "通知管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«NotificationPageVO»",
              "originalRef": "PageDTO«NotificationPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/admin/role/page": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "分页",
        "operationId": "queryPageUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "adminId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«RolePageVO»",
              "originalRef": "PageDTO«RolePageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/reconciliation/approve/detail/{id}": {
      "get": {
        "tags": [
          "对账管理"
        ],
        "summary": "queryApprovalReconciliationDetail",
        "operationId": "queryApprovalReconciliationDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApprovalDetailVO",
              "originalRef": "ApprovalDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/reconciliation/approve/{id}": {
      "put": {
        "tags": [
          "对账管理"
        ],
        "summary": "approveReconciliation",
        "operationId": "approveReconciliationUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "approvalDTO",
            "description": "approvalDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApprovalDTO",
              "originalRef": "ApprovalDTO"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/reconciliation/detail/{id}": {
      "get": {
        "tags": [
          "对账管理"
        ],
        "summary": "queryReconciliationDetail",
        "operationId": "queryReconciliationDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ReconciliationDetailVO",
              "originalRef": "ReconciliationDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/reconciliation/export": {
      "get": {
        "tags": [
          "对账管理"
        ],
        "summary": "导出对账数据报表",
        "operationId": "exportUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/reconciliation/page": {
      "get": {
        "tags": [
          "对账管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "meterNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "reconciliationNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "reconciliationStatus",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "userType",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«ReconciliationPageVO»",
              "originalRef": "PageDTO«ReconciliationPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApprovalDTO": {
      "type": "object",
      "properties": {
        "comment": {
          "type": "string",
          "refType": null,
          "example": "确认数据无误，予以批准"
        },
        "status": {
          "type": "string",
          "refType": null,
          "example": "APPROVED"
        }
      },
      "title": "ApprovalDTO"
    },
    "ApprovalDetailVO": {
      "type": "object",
      "properties": {
        "approvalRecordList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ApprovalRecordVO",
            "originalRef": "ApprovalRecordVO"
          },
          "example": [
            {
              "approvalOperator": "张三",
              "approvalStatus": "APPROVED",
              "approvalTime": "2023-05-15T14:30:00",
              "comment": "数据无误，批准通过",
              "reconciliationNo": 10023
            }
          ]
        },
        "balance": {
          "type": "number",
          "example": 450.75
        },
        "comment": {
          "type": "string",
          "example": "月度结算已完成"
        },
        "isApproved": {
          "type": "boolean",
          "example": true
        },
        "reconciliationNo": {
          "type": "integer",
          "format": "int64",
          "example": 10023
        },
        "status": {
          "type": "string",
          "example": "APPROVED"
        },
        "username": {
          "type": "string",
          "example": "李四"
        }
      },
      "title": "ApprovalDetailVO"
    },
    "ApprovalRecordVO": {
      "type": "object",
      "properties": {
        "approvalOperator": {
          "type": "string",
          "refType": null
        },
        "approvalStatus": {
          "type": "string",
          "refType": null
        },
        "approvalTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "comment": {
          "type": "string",
          "refType": null
        },
        "reconciliationNo": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "ApprovalRecordVO"
    },
...
    "LogPageVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "description": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "ip": {
          "type": "string",
          "refType": null
        },
        "module": {
          "type": "string",
          "refType": null
        },
        "operationType": {
          "type": "string",
          "refType": null
        },
        "operatorName": {
          "type": "string",
          "refType": null
        },
        "status": {
          "type": "string",
          "refType": null
        }
      },
      "title": "LogPageVO"
    },
...
    "NotificationDTO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "refType": null
        },
        "expireTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "notificationType": {
          "type": "string",
          "refType": null
        },
        "senderList": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "refType": "string"
        },
        "title": {
          "type": "string",
          "refType": null
        },
        "type": {
          "type": "string",
          "refType": null
        }
      },
      "title": "NotificationDTO"
    },
    "NotificationDetailVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "expireTime": {
          "type": "string",
          "format": "date-time"
        },
        "level": {
          "type": "string"
        },
        "senderName": {
          "type": "string"
        },
        "senderRole": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "NotificationDetailVO"
    },
    "NotificationPageVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "expireTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "level": {
          "type": "string",
          "refType": null
        },
        "readStatus": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "title": {
          "type": "string",
          "refType": null
        },
        "type": {
          "type": "string",
          "refType": null
        }
      },
      "title": "NotificationPageVO"
    },
...
    "PageDTO«LogPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LogPageVO",
            "originalRef": "LogPageVO"
          },
          "example": [
            {
              "createTime": "2023-06-10T08:25:13",
              "description": "更新用户信息",
              "id": 5012,
              "ip": "192.168.1.100",
              "module": "用户管理",
              "operationType": "UPDATE",
              "operatorName": "admin",
              "status": "成功"
            },
            {
              "createTime": "2023-06-10T08:20:45",
              "description": "创建新电表",
              "id": 5011,
              "ip": "192.168.1.100",
              "module": "电表管理",
              "operationType": "INSERT",
              "operatorName": "admin",
              "status": "成功"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 126
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 1256
        }
      },
      "title": "PageDTO«LogPageVO»"
    },
...
    "PageDTO«NotificationPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotificationPageVO",
            "originalRef": "NotificationPageVO"
          },
          "example": [
            {
              "content": "系统将于6月15日00:00-02:00进行升级维护，期间服务不可用",
              "createTime": "2023-06-10T10:00:00",
              "expireTime": "2023-06-15T02:00:00",
              "id": 1023,
              "level": "重要",
              "readStatus": 1,
              "title": "系统维护通知",
              "type": "系统通知"
            },
            {
              "content": "6月电费账单已生成，请及时缴纳",
              "createTime": "2023-06-05T00:00:00",
              "expireTime": "2023-06-25T23:59:59",
              "id": 1022,
              "level": "一般",
              "readStatus": 1,
              "title": "电费账单通知",
              "type": "账单通知"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 8
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 78
        }
      },
      "title": "PageDTO«NotificationPageVO»"
    },
...
    "PageDTO«ReconciliationPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReconciliationPageVO",
            "originalRef": "ReconciliationPageVO"
          },
          "example": [
            {
              "balance": 450.75,
              "reconciliationNo": 10023,
              "status": "已批准",
              "userId": 10025,
              "userType": "居民用户",
              "username": "张小明"
            },
            {
              "balance": 825.30,
              "reconciliationNo": 10022,
              "status": "已批准",
              "userId": 10026,
              "userType": "商业用户",
              "username": "李明"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 5
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 42
        }
      },
      "title": "PageDTO«ReconciliationPageVO»"
    },
    "PageDTO«RolePageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RolePageVO",
            "originalRef": "RolePageVO"
          },
          "example": [
            {
              "account": "admin",
              "adminId": 1001,
              "createTime": "2023-01-01T00:00:00",
              "role": "系统管理员",
              "roleDesc": "系统管理员，拥有所有权限",
              "status": 1
            },
            {
              "account": "operator",
              "adminId": 1002,
              "createTime": "2023-01-10T09:00:00",
              "role": "运营人员",
              "roleDesc": "负责日常运营工作",
              "status": 1
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 2
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 5
        }
      },
      "title": "PageDTO«RolePageVO»"
    },
...
    "R": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "业务状态码，200-成功，其它-失败"
        },
        "data": {
          "type": "object",
          "description": "响应数据"
        },
        "msg": {
          "type": "string",
          "example": "OK",
          "description": "响应消息"
        },
        "requestId": {
          "type": "string",
          "example": "1af123c11412e",
          "description": "请求id"
        }
      },
      "title": "R",
      "description": "通用响应结果"
    },
...
    "ReconciliationDetailVO": {
      "type": "object",
      "properties": {
        "approvalComment": {
          "type": "string"
        },
        "approvalOperator": {
          "type": "string"
        },
        "approvalTime": {
          "type": "string",
          "format": "date-time"
        },
        "balance": {
          "type": "number"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "meterNo": {
          "type": "string"
        },
        "paymentStatus": {
          "type": "string"
        },
        "reconciliationNo": {
          "type": "integer",
          "format": "int64"
        },
        "reconciliationStatus": {
          "type": "string"
        },
        "userPaymentRecordVOList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserPaymentRecordVO",
            "originalRef": "UserPaymentRecordVO"
          }
        },
        "userType": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "ReconciliationDetailVO"
    },
    "ReconciliationPageVO": {
      "type": "object",
      "properties": {
        "balance": {
          "type": "number",
          "refType": null
        },
        "meterNo": {
          "type": "string",
          "refType": null
        },
        "reconciliationNo": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "reconciliationStatus": {
          "type": "string",
          "refType": null
        },
        "reconciliationTime": {
          "type": "string",
          "format": "date",
          "refType": null
        },
        "userType": {
          "type": "string",
          "refType": null
        },
        "username": {
          "type": "string",
          "refType": null
        }
      },
      "title": "ReconciliationPageVO"
    },
...
    "RolePageVO": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "refType": null,
          "example": "admin"
        },
        "adminId": {
          "type": "integer",
          "format": "int64",
          "refType": null,
          "example": 1001
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-01-01T00:00:00"
        },
        "role": {
          "type": "string",
          "refType": null,
          "example": "系统管理员"
        },
        "roleDesc": {
          "type": "string",
          "refType": null,
          "example": "系统管理员，拥有所有权限"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "refType": null,
          "example": 1
        }
      },
      "title": "RolePageVO"
    },
...
    "UserPaymentRecordVO": {
      "type": "object",
      "properties": {
        "operator": {
          "type": "string",
          "refType": null,
          "example": "系统自动"
        },
        "paymentAmount": {
          "type": "number",
          "refType": null,
          "example": 138.30
        },
        "paymentMethod": {
          "type": "string",
          "refType": null,
          "example": "微信支付"
        },
        "paymentStatus": {
          "type": "string",
          "refType": null,
          "example": "成功"
        },
        "paymentTime": {
          "type": "string",
          "format": "date-time",
          "refType": null,
          "example": "2023-06-05T14:32:45"
        },
        "remark": {
          "type": "string",
          "refType": null,
          "example": "5月电费"
        }
      },
      "title": "UserPaymentRecordVO"
    },
...
  }
}